Focus Manager + Clipboard extractor
If the windows has the focus of the mouse, rightclicking inside the
input window will copy any string of the clipboard inside the input
field. if the windows does not have the focus, it must first get it in
order to extract the clipboard.

If the input field gets the focus and already has text inside it will
empty itself